+++ src/reporter/handler.rs
@@ -10,6 +10,7 @@ use storyteller::{EventHandler, Reporter};
 mod discard_output_handler;
 mod human_progress_handler;
 mod json_handler;
+mod minimal_output_handler;
 
 #[cfg(test)]
 mod testing;
@@ -17,6 +18,7 @@ mod testing;
 pub use discard_output_handler::DiscardOutputHandler;
 pub use human_progress_handler::HumanProgressHandler;
 pub use json_handler::JsonHandler;
+pub use minimal_output_handler::MinimalOutputHandler;
 
 #[cfg(test)]
 pub use testing::TestingHandler;